v6.5.3

+++ released/MyBox/src/main/java/mara/mybox/controller/BaseSysTableController.java
@@ -102,7 +102,7 @@ public abstract class BaseSysTableController<P> extends BaseTableViewController<
     }
 
     @Override
-    protected int clearData() {
+    protected long clearData() {
         if (tableDefinition != null) {
             return tableDefinition.deleteCondition(queryConditions);
         } else {
